fix(ccusage): address jsonl review feedback
Move the Claude byte_lines primitive into ccusage-jsonl so Claude and the other JSONL readers share the same crate without forcing Claude onto the slower marker scanner path.

Handle Codex and Qwen streaming read errors explicitly instead of treating them as EOF, keep Qwen streaming to avoid whole-file buffering, and remove the redundant OpenClaw marker guard identified by CodeRabbit and cubic.

Validation: cargo fmt --manifest-path rust/Cargo.toml --all; cargo test --manifest-path rust/Cargo.toml -p ccusage-jsonl; cargo test --manifest-path rust/Cargo.toml -p ccusage adapter::claude adapter::codex adapter::openclaw adapter::qwen via separate invocations; cargo build --manifest-path rust/Cargo.toml -p ccusage; cargo clippy --manifest-path rust/Cargo.toml --workspace --all-targets -- -D warnings; cargo build --manifest-path rust/Cargo.toml --release -p ccusage; bunx ccusage@20.0.5 daily/monthly --until 20260526 --offline --json matched the branch release output after jq -S; large-fixture hyperfine remained at parity/slightly faster.
